test(streaming): separate provider routing coverage
Move routing and failure-classification tests into their focused owner so the streaming service suite remains below the repository monolith threshold.
