+++ src/test/java/com/williamcallahan/javachat/service/OpenAIStreamingServiceTest.java
@@ -21,11 +21,6 @@ import com.openai.core.RequestOptions;
 import com.openai.core.http.Headers;
 import com.openai.core.http.StreamResponse;
 import com.openai.errors.InternalServerException;
-import com.openai.errors.NotFoundException;
-import com.openai.errors.OpenAIIoException;
-import com.openai.errors.PermissionDeniedException;
-import com.openai.errors.RateLimitException;
-import com.openai.errors.UnauthorizedException;
 import com.openai.models.ErrorObject;
 import com.openai.models.responses.ResponseCreateParams;
 import com.openai.models.responses.ResponseStreamEvent;
@@ -37,7 +32,6 @@ import com.williamcallahan.javachat.application.prompt.PromptTruncator;
 import com.williamcallahan.javachat.application.streaming.StreamingFailureReporter;
 import com.williamcallahan.javachat.domain.prompt.StructuredPrompt;
 import com.williamcallahan.javachat.web.SseConstants;
-import java.io.InterruptedIOException;
 import java.time.Duration;
 import java.util.List;
 import java.util.Objects;
@@ -48,17 +42,14 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.slf4j.LoggerFactory;
 import org.springframework.test.util.ReflectionTestUtils;
-import reactor.core.Exceptions;
 import reactor.core.publisher.Mono;
 import reactor.test.StepVerifier;
 
 /**
- * Verifies primary-provider backoff and streaming failure classification.
+ * Verifies streaming failure reporting, pre-text provider retries, and completion validation.
  *
- * <p>Backoff tests target {@link OpenAiProviderRoutingService#shouldBackoffPrimary} directly
- * (package-private, same package). Streaming recovery tests exercise the public
- * {@link OpenAIStreamingService#isRecoverableStreamingFailure} API that delegates to
- * the routing service.</p>
+ * <p>Provider routing policy is verified in {@link OpenAiProviderRoutingServiceTest} so this
+ * suite stays focused on SDK transport and stream-facing behavior.</p>
  */
 class OpenAIStreamingServiceTest {
     private final Logger serviceLogger = (Logger) LoggerFactory.getLogger(OpenAIStreamingService.class);
@@ -85,12 +76,6 @@ class OpenAIStreamingServiceTest {
         streamingFailureLogEvents.list.clear();
     }
 
-    private OpenAiProviderRoutingService createRoutingService() {
-        RateLimitService rateLimitService = mock(RateLimitService.class);
-        return new OpenAiProviderRoutingService(
-                rateLimitService, 600, RateLimitService.ApiProvider.GITHUB_MODELS.getName());
-    }
-
     private OpenAIStreamingService createStreamingService() {
         RateLimitService rateLimitService = mock(RateLimitService.class);
         OpenAiRequestFactory requestFactory =
@@ -99,112 +84,6 @@ class OpenAIStreamingServiceTest {
                 rateLimitService, 600, RateLimitService.ApiProvider.GITHUB_MODELS.getName());
         return new OpenAIStreamingService(
                 rateLimitService, requestFactory, providerRoutingService, new OpenAiStreamingFailureReporter());
-    }
-
-    @Test
-    void shouldBackoffPrimaryTreatsSdkIoAsBackoffEligible() {
-        OpenAiProviderRoutingService routingService = createRoutingService();
-        assertTrue(routingService.shouldBackoffPrimary(new OpenAIIoException("io")));
-    }
-
-    @Test
-    void shouldBackoffPrimaryIgnoresCallerCancellationWrappedBySdkIo() {
-        OpenAiProviderRoutingService routingService = createRoutingService();
-        InterruptedIOException interruptedRequest = new InterruptedIOException("request interrupted by caller timeout");
-        OpenAIIoException cancelledCompletion = new OpenAIIoException("Request failed", interruptedRequest);
-
-        assertFalse(routingService.shouldBackoffPrimary(cancelledCompletion));
-    }
-
-    @Test
-    void callerCancellationKeepsConfiguredPrimaryProviderEligible() {
-        RateLimitService rateLimitService = mock(RateLimitService.class);
-        when(rateLimitService.isProviderAvailable(RateLimitService.ApiProvider.OPENAI))
-                .thenReturn(true);
-        when(rateLimitService.isProviderAvailable(RateLimitService.ApiProvider.GITHUB_MODELS))
-                .thenReturn(true);
-        OpenAiProviderRoutingService routingService =
-                new OpenAiProviderRoutingService(rateLimitService, 600, RateLimitService.ApiProvider.OPENAI.getName());
-        OpenAIClient openAiClient = mock(OpenAIClient.class);
-        OpenAIClient githubModelsClient = mock(OpenAIClient.class);
-        InterruptedIOException interruptedRequest = new InterruptedIOException("request interrupted by caller timeout");
-        OpenAIIoException cancelledCompletion = new OpenAIIoException("Request failed", interruptedRequest);
-
-        routingService.recordProviderFailure(RateLimitService.ApiProvider.OPENAI, cancelledCompletion);
-
-        List<OpenAiProviderCandidate> availableProviders =
-                routingService.selectAvailableProviderCandidates(githubModelsClient, openAiClient);
-        assertEquals(2, availableProviders.size());
-        assertEquals(
-                RateLimitService.ApiProvider.OPENAI, availableProviders.get(0).provider());
-    }
-
-    @Test
-    void shouldBackoffPrimaryTreats401AsBackoffEligible() {
-        OpenAiProviderRoutingService routingService = createRoutingService();
-        Headers headers = Headers.builder().build();
-        UnauthorizedException unauthorized =
-                UnauthorizedException.builder().headers(headers).build();
-        assertTrue(routingService.shouldBackoffPrimary(unauthorized));
-    }
-
-    @Test
-    void shouldBackoffPrimaryTreats429AsBackoffEligible() {
-        OpenAiProviderRoutingService routingService = createRoutingService();
-        Headers headers = Headers.builder().build();
-        RateLimitException rateLimit =
-                RateLimitException.builder().headers(headers).build();
-        assertTrue(routingService.shouldBackoffPrimary(rateLimit));
-    }
-
-    @Test
-    void shouldBackoffPrimaryTreats404AsBackoffEligible() {
-        OpenAiProviderRoutingService routingService = createRoutingService();
-        Headers headers = Headers.builder().build();
-        NotFoundException notFoundException =
-                NotFoundException.builder().headers(headers).build();
-        assertTrue(routingService.shouldBackoffPrimary(notFoundException));
-    }
-
-    @Test
-    void shouldBackoffPrimaryDoesNotBackoffOnGenericRuntime() {
-        OpenAiProviderRoutingService routingService = createRoutingService();
-        assertFalse(routingService.shouldBackoffPrimary(new IllegalArgumentException("no")));
-    }
-
-    @Test
-    void recoverableStreamingFailureTreatsReactorOverflowTypeAsRetryable() {
-        OpenAIStreamingService streamingService = createStreamingService();
-        assertTrue(streamingService.isRecoverableStreamingFailure(Exceptions.failWithOverflow()));
-    }
-
-    @Test
-    void recoverableStreamingFailureTreatsValidationErrorsAsNonRetryable() {
-        OpenAIStreamingService streamingService = createStreamingService();
-        assertFalse(
-                streamingService.isRecoverableStreamingFailure(new IllegalArgumentException("bad request payload")));
-    }
-
-    @Test
-    void recoverableStreamingFailureTreatsNotFoundAsNonRetryable() {
-        OpenAIStreamingService streamingService = createStreamingService();
-        Headers headers = Headers.builder().build();
-        NotFoundException notFoundException =
-                NotFoundException.builder().headers(headers).build();
-        assertFalse(streamingService.isRecoverableStreamingFailure(notFoundException));
-    }
-
-    @Test
-    void recoverableStreamingFailureRejectsAuthenticationAuthorizationAndRateLimitFailures() {
-        OpenAIStreamingService streamingService = createStreamingService();
-        Headers headers = Headers.builder().build();
-
-        assertFalse(streamingService.isRecoverableStreamingFailure(
-                UnauthorizedException.builder().headers(headers).build()));
-        assertFalse(streamingService.isRecoverableStreamingFailure(
-                PermissionDeniedException.builder().headers(headers).build()));
-        assertFalse(streamingService.isRecoverableStreamingFailure(
-                RateLimitException.builder().headers(headers).build()));
     }
 
     @Test
@@ -422,48 +301,6 @@ class OpenAIStreamingServiceTest {
 
         verify(responseService).createStreaming(any(ResponseCreateParams.class), any(RequestOptions.class));
         verify(visibleTextDelta).delta();
-    }
-
-    @Test
-    void primaryBackoffDoesNotHideOnlyConfiguredProvider() {
-        RateLimitService rateLimitService = mock(RateLimitService.class);
-        when(rateLimitService.isProviderAvailable(RateLimitService.ApiProvider.OPENAI))
-                .thenReturn(true);
-        OpenAiProviderRoutingService routingService =
-                new OpenAiProviderRoutingService(rateLimitService, 600, RateLimitService.ApiProvider.OPENAI.getName());
-        OpenAIClient openAiClient = mock(OpenAIClient.class);
-
-        routingService.recordProviderFailure(RateLimitService.ApiProvider.OPENAI, new OpenAIIoException("io"));
-
-        List<OpenAiProviderCandidate> availableProviders =
-                routingService.selectAvailableProviderCandidates(null, openAiClient);
-
-        assertEquals(1, availableProviders.size());
-        assertEquals(
-                RateLimitService.ApiProvider.OPENAI, availableProviders.get(0).provider());
-    }
-
-    @Test
-    void primaryBackoffUsesConfiguredAlternateProviderWhenAvailable() {
-        RateLimitService rateLimitService = mock(RateLimitService.class);
-        when(rateLimitService.isProviderAvailable(RateLimitService.ApiProvider.OPENAI))
-                .thenReturn(true);
-        when(rateLimitService.isProviderAvailable(RateLimitService.ApiProvider.GITHUB_MODELS))
-                .thenReturn(true);
-        OpenAiProviderRoutingService routingService =
-                new OpenAiProviderRoutingService(rateLimitService, 600, RateLimitService.ApiProvider.OPENAI.getName());
-        OpenAIClient openAiClient = mock(OpenAIClient.class);
-        OpenAIClient githubModelsClient = mock(OpenAIClient.class);
-
-        routingService.recordProviderFailure(RateLimitService.ApiProvider.OPENAI, new OpenAIIoException("io"));
-
-        List<OpenAiProviderCandidate> availableProviders =
-                routingService.selectAvailableProviderCandidates(githubModelsClient, openAiClient);
-
-        assertEquals(1, availableProviders.size());
-        assertEquals(
-                RateLimitService.ApiProvider.GITHUB_MODELS,
-                availableProviders.get(0).provider());
     }
 
     @Test
